Create Add_new_attr_Value_by_using_3_strong_conditions.ps1
In my case, i need to change the city attr
if:
User have surname 0, and he is member of some group  in OU where his account stored. I Need to change his city to 0
or:
User have surname 1, and he is member of some group  in OU where his account stored. I Need to change his city to 1
else:
User surname not equal 0 or 1 and he isn`t member of list of groups.  I Need to change his city to 2

But this is universal scenario for tasks with similar requirements